2-Getting-Started/requirements.txt
langchain
python-dotenv
langchain-openai
langchain_community
ipykernel
beautifulsoup4
langchain_text_splitters
faiss-cpu
streamlit
langchain_ollama